yandex-jobs-bot/requirements.txt
python-telegram-bot==21.6
httpx==0.28.1
beautifulsoup4==4.12.3
apscheduler==3.10.4
python-dotenv==1.0.1
